Modify : NewProduct display(flex -> grid)
