+++ src/pages/Main/Main.scss
@@ -26,15 +26,10 @@
     .newProductsTitle {
       @include title();
     }
-
-    .newProducts {
-      display: flex;
-      margin-top: 30px;
-    }
   }
 
   .reviewContainer {
-    margin-top: 50px;
+    margin-top: 100px;
     width: 80%;
 
     .reviewTitle {
